fastapi/pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "Boilerplate code for a fastapi application"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.80.0"
uvicorn = "^0.18.2"
requests = "^2.28.1"
PyYAML = "^6.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
